+++ core/client/app/components/gh-ed-editor.js
@@ -32,7 +32,7 @@ export default TextArea.extend(EditorAPI, EditorShortcuts, EditorScroll, {
 
         this.setFocus();
 
-        this.sendAction('setEditor', this);
+        this.attrs.setEditor(this);
 
         run.scheduleOnce('afterRender', this, this.afterRenderEvent);
     },
@@ -41,22 +41,6 @@ export default TextArea.extend(EditorAPI, EditorShortcuts, EditorScroll, {
         if (this.get('focus') && this.get('focusCursorAtEnd')) {
             this.setSelection('end');
         }
-    },
-
-    /**
-     * Disable editing in the textarea (used while an upload is in progress)
-     */
-    disable() {
-        let textarea = this.get('element');
-        textarea.setAttribute('readonly', 'readonly');
-    },
-
-    /**
-     * Reenable editing in the textarea
-     */
-    enable() {
-        let textarea = this.get('element');
-        textarea.removeAttribute('readonly');
     },
 
     actions: {
