+++ core/client/app/components/gh-editor.js
@@ -1,20 +1,25 @@
 import Ember from 'ember';
+import ShortcutsMixin from 'ghost/mixins/shortcuts';
+import imageManager from 'ghost/utils/ed-image-manager';
+import editorShortcuts from 'ghost/utils/editor-shortcuts';
 
 const {Component, computed, run} = Ember;
 const {equal} = computed;
 
-export default Component.extend({
+export default Component.extend(ShortcutsMixin, {
     tagName: 'section',
-    classNames: ['gh-view'],
+    classNames: ['view-container', 'view-editor'],
 
+    activeTab: 'markdown',
+    editor: null,
+    editorDisabled: undefined,
+    editorScrollInfo: null, // updated when gh-ed-editor component scrolls
+    height: null, // updated when markdown is rendered
+    shouldFocusEditor: false,
     showCopyHTMLModal: false,
     copyHTMLModalContent: null,
 
-    // updated when gh-ed-editor component scrolls
-    editorScrollInfo: null,
-    // updated when markdown is rendered
-    height: null,
-    activeTab: 'markdown',
+    shortcuts: editorShortcuts,
 
     markdownActive: equal('activeTab', 'markdown'),
     previewActive: equal('activeTab', 'preview'),
@@ -25,8 +30,7 @@ export default Component.extend({
     // stays in sync
     scrollPosition: computed('editorScrollInfo', 'height', function () {
         let scrollInfo = this.get('editorScrollInfo');
-        let $previewContent = this.get('$previewContent');
-        let $previewViewPort = this.get('$previewViewPort');
+        let {$previewContent, $previewViewPort} = this;
 
         if (!scrollInfo || !$previewContent || !$previewViewPort) {
             return 0;
@@ -41,26 +45,74 @@ export default Component.extend({
         return previewPosition;
     }),
 
-    scheduleAfterRender() {
-        run.scheduleOnce('afterRender', this, this.afterRenderEvent);
-    },
-
     didInsertElement() {
         this._super(...arguments);
-        this.scheduleAfterRender();
+        this.registerShortcuts();
+        run.scheduleOnce('afterRender', this, this._cacheElements);
     },
 
-    afterRenderEvent() {
-        let $previewViewPort = this.$('.js-entry-preview-content');
+    willDestroyElement() {
+        if (this.attrs.onTeardown) {
+            this.attrs.onTeardown();
+        }
+        this.removeShortcuts();
+    },
 
+    _cacheElements() {
         // cache these elements for use in other methods
-        this.set('$previewViewPort', $previewViewPort);
-        this.set('$previewContent', this.$('.js-rendered-markdown'));
+        this.$previewViewPort = this.$('.js-entry-preview-content');
+        this.$previewContent = this.$('.js-rendered-markdown');
     },
 
     actions: {
         selectTab(tab) {
             this.set('activeTab', tab);
+        },
+
+        updateScrollInfo(scrollInfo) {
+            this.set('editorScrollInfo', scrollInfo);
+        },
+
+        updateHeight(height) {
+            this.set('height', height);
+        },
+
+        // set from a `sendAction` on the gh-ed-editor component,
+        // so that we get a reference for handling uploads.
+        setEditor(editor) {
+            this.set('editor', editor);
+        },
+
+        disableEditor() {
+            this.set('editorDisabled', true);
+        },
+
+        enableEditor() {
+            this.set('editorDisabled', undefined);
+        },
+
+        // The actual functionality is implemented in utils/ed-editor-shortcuts
+        editorShortcut(options) {
+            if (this.editor.$().is(':focus')) {
+                this.editor.shortcut(options.type);
+            }
+        },
+
+        // Match the uploaded file to a line in the editor, and update that line with a path reference
+        // ensuring that everything ends up in the correct place and format.
+        handleImgUpload(e, resultSrc) {
+            let editor = this.get('editor');
+            let editorValue = editor.getValue();
+            let replacement = imageManager.getSrcRange(editorValue, e.target);
+            let cursorPosition;
+
+            if (replacement) {
+                cursorPosition = replacement.start + resultSrc.length + 1;
+                if (replacement.needsParens) {
+                    resultSrc = `(${resultSrc})`;
+                }
+                editor.replaceSelection(resultSrc, replacement.start, replacement.end, cursorPosition);
+            }
         },
 
         toggleCopyHTMLModal(generatedHTML) {
