+++ core/client/app/mixins/editor-base-controller.js
@@ -1,7 +1,6 @@
 import Ember from 'ember';
 import PostModel from 'ghost/models/post';
 import boundOneWay from 'ghost/utils/bound-one-way';
-import imageManager from 'ghost/utils/ed-image-manager';
 
 const {Mixin, RSVP, computed, inject, observer, run} = Ember;
 const {alias} = computed;
@@ -17,7 +16,6 @@ PostModel.eachAttribute(function (name) {
 export default Mixin.create({
     _autoSaveId: null,
     _timedSaveId: null,
-    editor: null,
     submitting: false,
 
     showLeaveEditorModal: false,
@@ -126,7 +124,7 @@ export default Mixin.create({
             let markdown = model.get('markdown');
             let title = model.get('title');
             let titleScratch = model.get('titleScratch');
-            let scratch = this.get('editor').getValue();
+            let scratch = this.get('model.scratch');
             let changedAttributes;
 
             if (!this.tagNamesEqual()) {
@@ -253,31 +251,9 @@ export default Mixin.create({
     },
 
     actions: {
-        save(options) {
-            let prevStatus = this.get('model.status');
-            let isNew = this.get('model.isNew');
+        cancelTimers() {
             let autoSaveId = this._autoSaveId;
             let timedSaveId = this._timedSaveId;
-            let psmController = this.get('postSettingsMenuController');
-            let promise, status;
-
-            options = options || {};
-
-            // when navigating quickly between pages autoSave will occasionally
-            // try to run after the editor has been torn down so bail out here
-            // before we throw errors
-            if (!this.get('editor').$()) {
-                return 0;
-            }
-
-            this.toggleProperty('submitting');
-
-            if (options.backgroundSave) {
-                // do not allow a post's status to be set to published by a background save
-                status = 'draft';
-            } else {
-                status = this.get('willPublish') ? 'published' : 'draft';
-            }
 
             if (autoSaveId) {
                 run.cancel(autoSaveId);
@@ -288,10 +264,30 @@ export default Mixin.create({
                 run.cancel(timedSaveId);
                 this._timedSaveId = null;
             }
+        },
+
+        save(options) {
+            let prevStatus = this.get('model.status');
+            let isNew = this.get('model.isNew');
+            let psmController = this.get('postSettingsMenuController');
+            let promise, status;
+
+            options = options || {};
+
+            this.toggleProperty('submitting');
+
+            if (options.backgroundSave) {
+                // do not allow a post's status to be set to published by a background save
+                status = 'draft';
+            } else {
+                status = this.get('willPublish') ? 'published' : 'draft';
+            }
+
+            this.send('cancelTimers');
 
             // Set the properties that are indirected
             // set markdown equal to what's in the editor, minus the image markers.
-            this.set('model.markdown', this.get('editor').getValue());
+            this.set('model.markdown', this.get('model.scratch'));
             this.set('model.status', status);
 
             // Set a default title
@@ -345,51 +341,10 @@ export default Mixin.create({
             }
         },
 
-        // set from a `sendAction` on the gh-ed-editor component,
-        // so that we get a reference for handling uploads.
-        setEditor(editor) {
-            this.set('editor', editor);
-        },
-
-        // fired from the gh-ed-preview component when an image upload starts
-        disableEditor() {
-            this.get('editor').disable();
-        },
-
-        // fired from the gh-ed-preview component when an image upload finishes
-        enableEditor() {
-            this.get('editor').enable();
-        },
-
-        // Match the uploaded file to a line in the editor, and update that line with a path reference
-        // ensuring that everything ends up in the correct place and format.
-        handleImgUpload(e, resultSrc) {
-            let editor = this.get('editor');
-            let editorValue = editor.getValue();
-            let replacement = imageManager.getSrcRange(editorValue, e.target);
-            let cursorPosition;
-
-            if (replacement) {
-                cursorPosition = replacement.start + resultSrc.length + 1;
-                if (replacement.needsParens) {
-                    resultSrc = `(${resultSrc})`;
-                }
-                editor.replaceSelection(resultSrc, replacement.start, replacement.end, cursorPosition);
-            }
-        },
-
         autoSaveNew() {
             if (this.get('model.isNew')) {
                 this.send('save', {silent: true, backgroundSave: true});
             }
-        },
-
-        updateEditorScrollInfo(scrollInfo) {
-            this.set('editorScrollInfo', scrollInfo);
-        },
-
-        updateHeight(height) {
-            this.set('height', height);
         },
 
         toggleLeaveEditorModal(transition) {
