+++ core/client/app/mixins/editor-base-route.js
@@ -1,12 +1,18 @@
 import Ember from 'ember';
 import ShortcutsRoute from 'ghost/mixins/shortcuts-route';
 import styleBody from 'ghost/mixins/style-body';
-import editorShortcuts from 'ghost/utils/editor-shortcuts';
+import ctrlOrCmd from 'ghost/utils/ctrl-or-cmd';
 
 const {Mixin, RSVP, run} = Ember;
 
+let generalShortcuts = {};
+generalShortcuts[`${ctrlOrCmd}+alt+p`] = 'publish';
+generalShortcuts['alt+shift+z'] = 'toggleZenMode';
+
 export default Mixin.create(styleBody, ShortcutsRoute, {
     classNames: ['editor'],
+
+    shortcuts: generalShortcuts,
 
     actions: {
         save() {
@@ -22,14 +28,6 @@ export default Mixin.create(styleBody, ShortcutsRoute, {
 
         toggleZenMode() {
             Ember.$('body').toggleClass('zen');
-        },
-
-        // The actual functionality is implemented in utils/ed-editor-shortcuts
-        editorShortcut(options) {
-            // Only fire editor shortcuts when the editor has focus.
-            if (this.get('controller.editor').$().is(':focus')) {
-                this.get('controller.editor').shortcut(options.type);
-            }
         },
 
         willTransition(transition) {
@@ -96,8 +94,6 @@ export default Mixin.create(styleBody, ShortcutsRoute, {
             outlet: 'settings-menu'
         });
     },
-
-    shortcuts: editorShortcuts,
 
     attachModelHooks(controller, model) {
         // this will allow us to track when the model is saved and update the controller
