+++ core/client/app/utils/editor-shortcuts.js
@@ -1,13 +1,9 @@
 // # Editor shortcuts
-// Loaded by EditorBaseRoute, which is a shortcuts route
+// Loaded by gh-editor component
 // This map is used to ensure the right action is called by each shortcut
 import ctrlOrCmd from 'ghost/utils/ctrl-or-cmd';
 
 let shortcuts = {};
-
-// General editor shortcuts
-shortcuts[`${ctrlOrCmd}+alt+p`] = 'publish';
-shortcuts['alt+shift+z'] = 'toggleZenMode';
 
 // Markdown Shortcuts
