+++ core/client/tests/unit/components/gh-editor-test.js
@@ -9,9 +9,15 @@ describeComponent(
     'gh-editor',
     'Unit: Component: gh-editor',
     {
-        unit: true
+        unit: true,
         // specify the other units that are required for this test
-        // needs: ['component:foo', 'helper:bar']
+        needs: [
+            'component:gh-ed-editor',
+            'component:gh-ed-preview',
+            'helper:gh-count-words',
+            'helper:route-action',
+            'service:notifications'
+        ]
     },
     function () {
         it('renders', function () {
